Sentence-fill in emacs
When working with LaTeX, it is recommended to start each sentence on a
new line.  As an emacs user, I always want to make emacs do the
formatting for me.  This small project is to find a solution by looking
at the Emacs Lisp macros.  The single file "hack.el" in this project is
obtained by

  gunzip -c /opt/local/share/emacs/23.3/lisp/textmodes/fill.el.gz \
         > hack.el